requirements.txt
Django==2.2
django-ckeditor==6.0.0
django-environ==0.4.5
django-modeltranslation==0.16
django-phonenumber-field==5.0.0
gunicorn==20.0.4
phonenumberslite==8.12.12
Pillow==8.0.1
psycopg2==2.8.6
django-admin-sortable2==0.7.7
pytils==0.3
django-colorfield==0.3.2
solid-i18n==1.4.2
